Factor out program args into its own file
