+++ tipper.py
@@ -1,5 +1,5 @@
-import argparse
 from emailSender import send_email
+from programArgs import getArgs
 import requests
 from datetime import datetime, timedelta
 import time
@@ -16,10 +16,7 @@ MIN_SPOTS = 2
 EMAIL_REGEX = r'\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b'
 
 def main():
-    arg_parser = init_argparse()
-    args = arg_parser.parse_args()
-
-    password = args.pwd
+    password = getArgs().pwd
     requests_session = requests.session()
 
     email_recipients = getInputEmails()
@@ -29,7 +26,7 @@ def main():
         1501796650: "Twilight",
         1501386657: "Front 10"
     }
-    moorePark = GolfCourse("Moore Park", "https://moorepark.miclub.com.au/guests/bookings/ViewPublicTimesheet.msp?bookingResourceId=3050007", mooreParkFeeGroups, args.no_cache)
+    moorePark = GolfCourse("Moore Park", "https://moorepark.miclub.com.au/guests/bookings/ViewPublicTimesheet.msp?bookingResourceId=3050007", mooreParkFeeGroups)
 
     eastLakeFeeGroups = {
         10230951: "Mon - Fri (before 1pm)",
@@ -38,7 +35,7 @@ def main():
         2251832: "Weekend (1pm - 2pm)",
         10230962: "Sundowner (after 2pm)"
     }
-    eastLake = GolfCourse("East Lake", "https://www.eastlakegolfclub.com.au/guests/bookings/ViewPublicTimesheet.msp?bookingResourceId=3000000", eastLakeFeeGroups, args.no_cache)
+    eastLake = GolfCourse("East Lake", "https://www.eastlakegolfclub.com.au/guests/bookings/ViewPublicTimesheet.msp?bookingResourceId=3000000", eastLakeFeeGroups)
 
     golfCourses = [moorePark, eastLake]
     print()
@@ -56,23 +53,10 @@ def main():
             new_tee_times_by_course[golfCourse.name] = new_tee_times
 
     if len(new_tee_times_by_course) != 0:
-        send_email(email_recipients, new_tee_times_by_course, password, args.local)
+        send_email(email_recipients, new_tee_times_by_course, password)
     t1 = time.time()
     print(f'Scrape took {t1-t0} seconds')
     print()
-
-def init_argparse() -> argparse.ArgumentParser:
-    parser = argparse.ArgumentParser()
-
-    parser.add_argument('-p','--pwd', help="Password for sending tipper email.")
-    parser.add_argument("email_recipients", nargs='*');
-
-    # Dev flags
-    parser.add_argument(
-        "-l", "--local", action="store_true", help="If enabled, prints email to terminal instead of sending.")
-    parser.add_argument("--no_cache",  action="store_true", help="If enabled, ignores cache of perviously seen times when generating tee times")
-
-    return parser
 
 def getInputEmails():
     email_recipients = []
